+++ modules/@apostrophecms/image/ui/apos/apps/AposImageRelationshipQueryFilter.js
@@ -1,5 +1,5 @@
 export default () => {
-  const queryOptions = [ 'minSize' ];
+  const queryOptions = [ 'minSize', 'aspectRatio' ];
 
   apos.bus.$on('piece-relationship-query', (query) => {
     const [ options = {} ] = apos.area.widgetOptions || [];